% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combining}
\alias{combining}
\alias{combining.default}
\alias{combining_mean}
\alias{combining_sd}
\alias{combining.data.frame}
\title{Combining Means and SD by groups}
\usage{
combining(...)

\method{combining}{default}(m1, m2, sd1 = 1, sd2 = 1, n1 = 5, n2 = 5)

combining_mean(x1, x2, n1 = 5, n2 = 5)

combining_sd(m1, m2, sd1, sd2, n1 = 5, n2 = 5)

\method{combining}{data.frame}(
  x,
  n = c("n", "cn", "tn", "cn", "tn", "cn", "tn"),
  m = c("m", "cm", "tm", "tm1", "tm2", "cm1", "cm2"),
  sd = c("sd", "csd", "tsd", "tsd1", "tsd2", "csd1", "csd2"),
  ...
)
}
\arguments{
\item{...}{}

\item{m1, m2}{Mittelwerte}

\item{sd1, sd2}{Standartabweichung}

\item{n1, n2}{Stichprobe}

\item{x}{data.frame (2 Rows)}

\item{n}{Names  of Number}

\item{m}{Names  of Estimated mean}

\item{sd}{Names  of Standard deviation}
}
\value{
vector(m, sd, n)
}
\description{
Combining Means and SD by groups
}
\examples{

combining(56, 50,2.3, 5.1,3	, 19)
combining(55.2, 50.4,2.2,	5.2,3	, 19)
combining(53.4,  49.3, 9.1,	8.7, 13,	26)
combining( 53.5, 49.4, 8.6,	8.4,13,	26)
set.seed(1)

n <- 20
dat <- data.frame(x = c(rnorm(n, 50, 5), rnorm(n, 40, 8)),
                  group = gl(2, n, labels = c("Control", "Treat")))[-c(1:5), ]

my_m <- function(x)
  round(c(m = mean(x), sd = sd(x), n = length(x)), 1)
aggregate(x ~ group, dat, FUN = my_m)

my_m(dat$x)


round(combining(51.1, 40.0, 4.7, 6.1, 15, 20), 1)
 dat <-
data.frame(
  #author= c("A", "B"),
  cn = c(134, 38),
  cm = c(5.96, 5.34),
  csd = c(4.24, 4),

  tn = c(113, 35),
  tm = c(6.82, 6.12),
  tsd = c(4.72, 4.32)

)


combining(dat)

}
