% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{convert.default}
\alias{convert.data.frame}
\title{Converting Among Effect Sizes}
\usage{
convert(...)

\method{convert}{default}(
  cn = 5,
  cm = NA,
  csd = NA,
  cse = NA,
  cm1 = NA,
  csd1 = NA,
  cse1 = NA,
  cm2 = NA,
  csd2 = NA,
  cse2 = NA,
  tn = 5,
  tm = NA,
  tsd = NA,
  tse = NA,
  tm1 = NA,
  tsd1 = NA,
  tse1 = NA,
  tm2 = NA,
  tsd2 = NA,
  tse2 = NA,
  r = NA,
  md = NA,
  sd = NA,
  se = NA,
  d = NA,
  g = NA,
  f = NA,
  digits = 2,
  ...
)

\method{convert}{data.frame}(x, digits = 3, ...)
}
\arguments{
\item{...}{not used}

\item{cn}{Number of observations in control group.}

\item{cm}{Estimated mean in control group.}

\item{csd, cse}{Standard deviation in control group.}

\item{cm1, csd1, cse1, cm2, csd2, cse2}{control group}

\item{tn}{Number of observations in experimental group.}

\item{tm}{Estimated mean in experimental group.}

\item{tm1, tsd1, tse1, tm2tsd2, tse2, , }{experimental group.}

\item{r, d, g, f}{standardisierte Effecte}

\item{md, sd, se}{Effecte}

\item{digits}{digits}

\item{tsdtse, }{Standard deviation in experimental group.}
}
\value{
Vector

data.frame
}
\description{
Quelle:
 Sutton seite 29 formel 2.16
 calculating missing SDs
Morris, Scott B., und Richard P. DeShon. „Combining Effect Size Estimates in Meta-Analysis with Repeated Measures and Independent-Groups Designs.“ Psychological Methods, Bd. 7, Nr. 1, 2002, S. 105–25.
DOI.org (Crossref), https://doi.org/10.1037/1082-989X.7.1.105.

change from baseline (also called a change score)
mean difference (or difference in means)
}
\details{
https://training.cochrane.org/handbook/current/chapter-06
}
