% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_extract.R
\name{extract}
\alias{extract}
\alias{tbll_extract.meta}
\title{Extract Meta}
\usage{
tbll_extract.meta(
  x,
  digits = 2,
  include.I2 = TRUE,
  include.Q = TRUE,
  include.tau = TRUE,
  include.H = TRUE,
  ...
)
}
\description{
tbll_extract.meta: Ausgabe von print.meta als tabelle
}
\examples{
# library(meta)
data(Olkin1995)
head(Olkin1995)
m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = c(41, 47, 51, 59),
              studlab = paste(author, year),
              method = "Inverse")
print(m1, digits = 1, digits.Q = 1)
Tbll(m1)
}
