% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converting-Among-Effect-Sizes.r
\name{from_ci}
\alias{from_ci}
\alias{sd_from_ci}
\alias{sd_from_se}
\alias{se_from_sd}
\alias{se_from_ci}
\alias{se_from_d}
\alias{ci_from_d}
\title{Converting Among Effect Sizes}
\usage{
sd_from_ci(
  ci_low,
  ci_hig,
  n = n1 + n2,
  n1 = 5,
  n2 = 5,
  ci = 0.95,
  verbose = TRUE
)

sd_from_se(se, n = n1 + n2, n1 = 5, n2 = 5, verbose = FALSE)

se_from_sd(sd, n = n1 + n2, n1 = 5, n2 = 5, verbose = TRUE)

se_from_ci(
  ci_low,
  ci_hig,
  n = n1 + n2,
  n1 = 5,
  n2 = 5,
  ci = 0.95,
  verbose = TRUE
)

se_from_d(d, n1, n2)

ci_from_d(d_obs, n1, n2)
}
\arguments{
\item{ci_low}{lower limit}

\item{ci_hig}{upper limit}

\item{n}{sample size}

\item{ci}{confidence interva}
}
\value{
value
}
\description{
Confidence interval based on computation
https://stats.stackexchange.com/questions/495015/what-is-the-formula-for-the-standard-error-of-cohens-d
}
